separate action and execution
